- Moved `BirthAnnouncement` from `procreation` package to `lifeEvents` for better organization.
